Move magic/chunk into separate file
